Simple bash script to notify when new email has arrived.
This is intended to be used with offlineimap (with or without mutt).

You can specify which mailbox directories to monitor for new mail,
and the mailbox dirs are efficiently monitored with inotifywait.

Events monitored are 'create' and 'moved_to', which should work for
offlineimap.

Notifications are sent via notify-send (part of libnotify), which should
work with XFCE, KDE, GNOME 2/3, Unity, etc..

You may need to install the inotify-tools package for inotifywait.